New securityContext options and annotations (#131)
Adds an option for running the injected vault containers as the
User of the first application container in the Pod. Requires
`Spec.Containers[0].SecurityContext.RunAsUser` to be set in the
pod spec. Defaults to false. Available as:

- `-run-as-same-user` command-line option
- `AGENT_INJECT_RUN_AS_SAME_USER` environment variable
- `vault.hashicorp.com/agent-run-as-same-user` annotation

Also adds an option for controlling whether or not the SecurityContext
is set on injected containers. Defaults to true. Available as:

- `-set-security-context` command-line option
- `AGENT_INJECT_SET_SECURITY_CONTEXT` environment variable
- `vault.hashicorp.com/agent-set-security-context` annotation

Co-authored-by: Mattia <mmascia@redhat.com>