Add file annotation (#158)
Adds new annotation `vault.hashicorp.com/agent-inject-file`. This sets
the filename and path in the secrets volume where a vault secret will
be written. The container mount path for the secrets volume may be
modified with the `secret-volume-path` annotation.

This allows for filenames that aren't limited in length as k8s
annotation keys, and for arbitrary directory structures in the secrets
volume.

Co-authored-by: Samuel Beaulieu <samuel@puppetlabs.com>