add agent-enable-quit annotation (#330)
New annotation to enable the /agent/v1/quit endpoint on an
injected agent. This option defaults to false, and will be set if
true on the existing cache listener, or a new localhost listener
with a basic cache stanza configured since that's required for an
agent listener.

Also add `omitempty` to `use_auto_auth_token`.

Only set `enable_quit` if the annotation is true. Since the agent
is defaulting to false, and it's very unlikely that will
change. This also meant the Agent.EnableQuit field doesn't need
to be a pointer any longer, so refactored that to make the logic
a little simpler.