Switch to newer json-patch; remove dead code
Switch from `github.com/mattbaird/jsonpatch` (unmaintained) to a new
version of `github.com/evanphx/json-patch`.

I also noticed that the `agent.Patches` object is unused, as is the
`patches` parameter of `agent.New()`, so I removed those, which
simplifies some of the code.

(This is in preparation for some more work involving JSON patches,
so I wanted to clean up existing usage of it.)