Set Kubernetes user-agent to include version (#411)
We weren't setting the user-agent, so Kubernetes sets a default
of the command-line (`vault-k8s`, which is okay) and an automatically
set version in the Kubernetes code base. This is probably not what
we want, since it implies that the Kubernetes version is the
`vault-k8s` version, and it is often just `0.0.0`.
See https://github.com/kubernetes/client-go/blob/master/pkg/version/base.go

Instead, we set the user-agent based on our own version.

I didn't see an easy way to test this in a unit test, but I did
test it manually by running `vault-k8s` with the environment
variable `GODEBUG=http2debug=1` set.