Don't override shareProcessNamespace unless annotation is explicitly set
By having a default of `false`, we would automatically set
`shareProcessNamespace` to `false` when injecting the agent pod,
regardless of what the main manifest had set.

Instead, we should only set `shareProcessNamespace` if the annotation
is explicitly set, and otherwise let the default from the main manifest
stand.

Fixes #436 (thanks @justinas-b for reporting it).

I tested this by:

* Before this patch:
  * Create a pod with `shareProcessNamespace: true` with injected agent
  * Verifying that the containers cannot see each other with `ps -ax`
* After this patch:
  * Create a pod with `shareProcessNamespace: true` with injected agent and no annotation
  * Verifying that the containers *can* see each other with `ps -ax`
  * Create a pod with `shareProcessNamespace: true` with injected agent and annotation set to `'false'`
  * Verifying that the containers *cannot* see each other with `ps -ax`